server: proxy: refactor

+++ server/proxy/modules/modules_api.h
@@ -106,7 +106,7 @@ typedef struct channel_data_event_info
 
 	/* actual data */
 	const BYTE* data;
-	int data_len;
+	size_t data_len;
 } proxyChannelDataEventInfo;
 #define WINPR_PACK_POP
 #include <winpr/pack.h>
